Move flatten_dict familly to dataset utils
